Unit tests for ViewState
